Add useUserPreferences.ts
